+++ ui/src/shared/middleware/resizeLayout.js
@@ -1,4 +1,6 @@
 // Trigger resize event to relayout the React Layout plugin
+import queryString from 'query-string'
+import {enablePresentationMode} from 'src/shared/actions/app'
 
 export default function resizeLayout() {
   return next => action => {
@@ -11,6 +13,13 @@ export default function resizeLayout() {
       const evt = document.createEvent('HTMLEvents')
       evt.initEvent('resize', false, true)
       window.dispatchEvent(evt)
+    }
+
+    const qs = queryString.parse(window.location.search)
+
+    if (qs.present === 'true') {
+      next(enablePresentationMode())
+      next(action)
     }
   }
 }